requirements.txt
streamlit
google-generativeai
pandas
graphviz